Enable zone-aware fallback demo
To see the customers fallback to another zone you need two sets of stores
services with different domains in their hostname (e.g. cfapps.io and
cf-apps.com). The domain is used as a "zone" and preference is given to the
stores in the same zone as the customers. The customers zone is defined by
a config entry (DOMAIN or APPLICATION_DOMAIN so the CF scripts here will
work).